feature/DAILY-02 - Continuidade no refactor do backend, separação em novos packages
