+++ src/main/java/com/daily/config/authentication/dto/AuthenticationDTO.java
@@ -1,4 +1,4 @@
-package com.daily.dto;
+package com.daily.config.authentication.dto;
 
 public record AuthenticationDTO(String login, String password) {
 }
