+++ src/main/java/com/daily/config/infra/AuthorizationService.java
@@ -1,11 +1,12 @@
-package com.daily.services;
+package com.daily.config.infra;
 
-import com.daily.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import com.daily.user.repository.UserRepository;
 
 @Service
 public class AuthorizationService implements UserDetailsService {
