+++ src/main/java/com/daily/config/infra/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.daily.config;
+package com.daily.config.infra;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
