+++ src/main/java/com/daily/config/infra/SecurityConfigurations.java
@@ -1,4 +1,4 @@
-package com.daily.infra.security;
+package com.daily.config.infra;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
