+++ src/main/java/com/daily/config/infra/TokenService.java
@@ -1,10 +1,11 @@
-package com.daily.infra.security;
+package com.daily.config.infra;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTCreationException;
 import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.daily.model.User;
+import com.daily.user.model.User;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
