--- src/main/java/com/daily/dto/TaskRequestDTO.java
@@ -1,5 +1,0 @@
-package com.daily.dto;
-
-import com.daily.model.User;
-
-public record TaskRequestDTO(String description) {}
