+++ src/main/java/com/daily/task/dto/TaskRequestDTO.java
@@ -1,0 +1,5 @@
+package com.daily.task.dto;
+
+import com.daily.user.model.User;
+
+public record TaskRequestDTO(String description) {}
