+++ src/main/java/com/daily/user/dto/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.daily.dto;
+package com.daily.user.dto;
 
 public record LoginResponseDTO(String token) {
 }
