+++ src/main/java/com/daily/user/dto/RegisterDTO.java
@@ -1,6 +1,6 @@
-package com.daily.dto;
+package com.daily.user.dto;
 
-import com.daily.model.UserRole;
+import com.daily.user.model.UserRole;
 
 public record RegisterDTO(String login, String password, UserRole role) {
 }
