+++ src/main/java/com/daily/user/repository/UserRepository.java
@@ -1,8 +1,9 @@
-package com.daily.repository;
+package com.daily.user.repository;
 
-import com.daily.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
+
+import com.daily.user.model.User;
 
 public interface UserRepository extends JpaRepository<User, String> {
     UserDetails findByLogin(String login);
